+++ projectEulerProblems/src/Problem44.java
@@ -1,12 +1,12 @@
 
 public class Problem44 {
-	public static long getPentagonal(long n) {
+	public static int getPentagonal(int n) {
 		return (3 * n - 1) * n / 2; 
 	}
 	
-	public static Boolean isPentagonal(long n) {
-		long m = 1;
-		long k = 1;
+	public static Boolean isPentagonal(int n) {
+		int m = 1;
+		int k = 1;
 		while (n >= k) {
 			if (n == k) return true;
 			k += 3 * m + 1;
@@ -15,11 +15,11 @@ public class Problem44 {
 		return false;
 	}
 	
-	public static long solve () {
-		long a = 1;
-		for (long m = 1; m < 1000000; m++) {
-			long b = 1;
-			for (long n = 1; n < m; n++) {
+	public static int solve () {
+		int a = 1;
+		for (int m = 1; m < 1000000; m++) {
+			int b = 1;
+			for (int n = 1; n < m; n++) {
 				if (isPentagonal(a + b)) {
 					if (isPentagonal(a + 2 * b)) return a;
 					if (isPentagonal(2 * a + b)) return b;
